changing filenames to snake_case
