+++ include/mppic/critic_function.hpp
